[CPDLP-1689] Refactor delivery partner show controller action
